Values never change
The keep scroll function almost works. Just trying to figgure out why it doesn't change when I scroll anymore. Think it might be window.pageYOffset making problems. I encountered issues with that one earlier in testing. If window.pageYOffset is set to a value, it will not change anymore when you scroll. I hope changing the order  of the selected values for the variable "top" will be an easy fix.