Wierd scroll value behaviour
The scrollTop value jumps to 100 after scrolling back to the top. Seems like an easy enough fix though, since it only does this from a value under 10. Simpy fixed it by checking if the previous scroll value was under 10, and the current is 100. If That is the case, it resets to 0