Performance update and git ignore
Trying out something new with the buttonController.py script. It uses 10% of the cpu when the button is in a state of noise. For this reason, I am trying out a new way of noise reduction for my button. If you are planning on creating a button for your lights with my system, I recomend making your own script, as buttonController.py is made to work with my makeshift button, and therefore has to do a lot of error correction because of a bad button. Also made sure the node_modules folder is not uploaded to github, since there is no reason for it to be.