Inverting speed value and error handling
Forgot about the fact that getDataval() will return False if it is unable to open the JSON file at the moment. Made sure that the theaterChase function is compatible with this, and also made sure to invert the speed variable by subtracting it from 100. This way the slider on the control panel is the right way.