Checking if something works
A piece of code that makes the colorDrip function turn off the LEDs after turning on the next in the row makes no sense to me anymore. I don't remember how it works, or why I made it. Soo, I am trying out a oneliner that in theory should do the same thing. Only one way to find out.