Removing autorun suppport
Using autorun when the script is still in beta souds like a bad idea now that I think about it. For some reason the script can't find what it is looking for when I try to start it otherwise, so I reverted back to how it was before I made it autorun compatible.