confusing error
Trying out something as random as the error. I mean, I would have to do this change anyway, but let's hope my string starts being defined as well.
WHAT THE HELL does even this mean?:
name "noiseReduce" is not defined.
That is the string saved in the variable mode. What the fuck?