Activate in miliseconds
This should make the script calculate how many milliseconds there are until the next beat in the song the user is currently listening to on spotify. With this data I will make the python script sync up with the beat in the song perfectly instead of leaving it up to chance.