3 most average
Trying to limit stuttering in cases where the monitor color scanner has to deal with images that have very close amounts of pixels per color. In this case, the script would previously jump between the colors often, and never settle for a color. By getting the average of the top three, I am hoping to get less of that stuttering.