Removed old code, and exposed settings
I removed the old code that found the most frequent color, as that is no longer in use. Also exposed the "step" variable that changes the speed and presicion of the script. This allows for custom speed and presicion for computers so the script can run smoothly on any computer.