Proper spring cleaning
Ok, this is a big one. Only visual changes on the website, and a lot of old code removal. Everything I could find that has been made irrelevant by other systems have been removed, and I split the control panel into 3 pages. This should make things load faster, and look better. Also changed some of the javascript to be compatible with the change in website layout