New web controls for spotify sync
The website is now able to talk to the server, and initiate a BPM request to spotify. The same button will also redirect the client to a login page if no data is recieved from spotify.

It is important to note that the server will send a 400 bad response in two cases:
1. The server is not logged in, and a simple login is required.
2. There is no song playing on the currently logged in account