Check for no spotify sync
The songDone value in the bpm.json file defaults to 0 when there is no song synced, because changing the BPM for the BPM mode will set this value to 0. This addition to the code makes sure lightController.py understands what this means, and doesn't try to get new song data from spotify.

I also made sure that the code automatically resets doneAt when no data is recieved when lightController.py asks for it. This prevents the code from stopping when the client stops listening to music, or catches any other errors that get through