Set timeout for request
The code was looping forever, and then crashing because it kept asking for the same thing again and again without getting it. Added a timeout to the code, so it stops after 1 second of this.

Not very worried about this desyncing songs, as this part of the code is only run when a new song has started, or when the function is activated.